Ui text: fix deletion of swallowed object
The internal Interactive.Text object is a composite object. Its parent
is set to be this object, too.

However, it's also being swallowed in the edje object of this widget.
The problem is that upon getting swallowed, the parent is reset to be
the edje object, and when the edje object gets deleted (destroyed along
with the current widget, the swallowed object's parent get reparented to
the canvas (???).
So, adding this fix/hack. Not sure of the nature of this form of
handling.