evas map: Get rid of UV coordinates (internally)
Honestly... I don't understand what they were for. I can't figure
out why uv.{w,h} should be different from image.{w.h} and in which
situation it made sense.

In fact in the evas example the values would end up invalid (too
large, until the previous patch). The actual output also looked
wrong as a result.

This simplifies the code a bit and avoids making mistakes like the
one I'm trying to fix in the previous patch. Now the image size
of both the proxy and its source (if it's also an evas image) should
be the same, so the math will work the same.

Note: This does not remove the image UV values set on the map.

@fix